[LUCENENET-409] Invalid Base exception in DateField.StringToTime()
git-svn-id: https://svn.apache.org/repos/asf/incubator/lucene.net@1094726 13f79535-47bb-0310-9956-ffa450edef68

+++ trunk/C#/src/Lucene.Net/SupportClass.cs
@@ -699,7 +699,7 @@ public class SupportClass
         public static long ToInt64(System.String s)
         {
             long number = 0;
-            int factor;
+            long factor;
 
             // handle negative number
             if (s.StartsWith("-"))
